Add typescript compilation to Makefile, and start splitting things into modules
