+++ Makefile
@@ -1,5 +1,6 @@
 .PHONY: build-arm
 build-arm: install-packr
+	tsc -p ./webserver
 	GOOS=linux GOARCH=arm GOARM=7 packr build -ldflags="-s -w" ./cmd/feverscreen
 
 .PHONY: install-packr
@@ -8,6 +9,7 @@ install-packr:
 
 .PHONY: build
 build: install-packr
+	tsc -p ./webserver
 	packr build -ldflags="-s -w" ./cmd/feverscreen
 
 .PHONY: release
@@ -18,3 +20,7 @@ release: install-packr
 clean:
 	packr clean
 	rm feverscreen
+
+.PHONY: install-typescript
+install-typescript:
+	npm install -g typescript
